Updated configurations and tensorflow

+++ docker.sh
@@ -8,10 +8,15 @@
 disp "Docker REINSTALLATION Script"
 #logging/utils/help
 dry_echo=""
+AUTO=0
 if test "$1" = "--dry-run" -o "$1" = "-D" ; then 
 	echo "Running docker setup in --dry-run mode"
-	#set -v
 	dry_echo="echo "
+	shift
+fi
+if test "$1" = "--automated" -o "$1" = "-a" ; then 
+	echo "Running docker setup in automated mode"
+	AUTO=1
 fi
 ## If you are running docker directly from here, please set the SUDO choice here.
 if [ -z $Install_Docker ]; then
@@ -19,29 +24,33 @@ if [ -z $Install_Docker ]; then
 	
 	## This is the variable to set - if you set this to 1 you don't have to "sudo docker" everytime
 	export Docker_Remove_SUDO=0
-
-	read -p "Press Enter to exit, or any other key to continue: - " exitqn
-	if test "$exitqn" = ""; then
-		echo "Exiting"; exit 127
-	else
-		echo "Continuing Docker Installation"
-	fi
-	
-	read -p "Uninstall previous versions of Docker? (y/Y for yes) - " uninstqn
-	if test "$uninstqn" = "N" -o "$uninstqn" = "n"; then
-		echo "Not uninstalling previous versions of Docker.";
-	else
-		echo "Uninstalling previous versions of Docker.";
-		sudo apt-get remove -y docker docker-engine docker.io
+	if [ $AUTO -eq 0 ]; then
+		read -p "Press Enter to exit, or any other key to continue: - " exitqn
+		if test "$exitqn" = ""; then
+			echo "Exiting"; exit 127
+		else
+			echo "Continuing Docker Installation"
+		fi
+		
+		read -p "Uninstall previous versions of Docker? (y/Y for yes) - " uninstqn
+		if test "$uninstqn" = "N" -o "$uninstqn" = "n"; then
+			echo "Not uninstalling previous versions of Docker.";
+		else
+			echo "Uninstalling previous versions of Docker.";
+			sudo apt-get remove -y docker docker-engine docker.io
+		fi
 	fi
 fi
 
 #sudo apt-get install -y linux-image-extra-$(uname -r) linux-image-extra-virtual}"
 #echo "Installing docker dependencies";
 $dry_echo sudo apt-get install -y apt-transport-https ca-certificates curl software-properties-common;
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+if [ "$dry_echo" = "" ]; then 
+	curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
+fi
 echo "Adding docker repository to lists";
 $dry_echo sudo add-apt-repository -y "deb [arch=amd64] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" # &> /dev/null
+#$dry_echo sudo add-apt-repository "deb https://apt.dockerproject.org/repo/ ubuntu-$(lsb_release -cs) main"
 
 #echo "Running apt-get update"; 
 $dry_echo sudo apt-get update
@@ -57,11 +66,13 @@ if [ $Docker_Remove_SUDO -eq 1 ]; then
 else
 	$dry_echo sudo docker run hello-world
 	# Verifying if sudo docker run... works.
-	if [  $? -eq 0 ]; then
-		log "Docker installed fine"
-		disp "Docker installed fine!"
-	else
-		log "Docker did not install fine"
-		disp "Docker did not install fine!"
+	if [ "$dry_echo" = "" ]; then 
+		if [  $? -eq 0 ]; then
+			log "Docker installed fine"
+			disp "Docker installed fine!"
+		else
+			log "Docker did not install fine"
+			disp "Docker did not install fine!"
+		fi
 	fi
 fi
